Change section on URL shortening to use bit.ly
All of the current URL shortner gems are busted in some way. urlshortner is DOA along with RubyURL.com and doesn't run in 1.9. The bit.ly gem/API require API/login credentials. The tinyurl gem is old and broken. The googl gem has an implicit dependency on ActiveSupport and breaks outside of it.