Section Tag
Created a section tag which allows you to specify a page that will be
displayed inline, an image, and the remaining text makes up the Title.

This is currently only in use on the main tutorial index to make it more
attractive to the various areas of our site.

NOTE: This commit does not address issues with the correct linking of
the javascript content.