Refer to Sublime Text rather than RubyMine
Sublime Text 2 has replaced RubyMine as the simplest thing that could
possibly work across all platforms.

This means that ruby programs are run directly in the terminal rather
than in RubyMine's console.

* the Blogger project
* the EventManager project
* the Ruby in 100 Minutes tutorial

The RubyMine references have been left alone in the Debugging tutorial,
because it is clearly marked as being a rare and curious thing.

Also, it is left in the Environment section, as it is still a valid
choice of editor/IDE.