Introduce 1.9 hash syntax (100 minutes)
I added a small section that explains the 1.9 syntax after the
traditional style is introduced, rather than starting with it.

This is because the whole "It's a symbol, even though it doesn't have a
colon in front of it" is a bit easier to explain when it isn't mixed
into the actual explanation for hash.