Attempt to clarify requirements for EventReporter
One of the students mentioned that they didn't understand the bit about
the commands for the program, so I've tried to clarify.

Also, it seems like a bunch of students didn't actually read through the
test scenarios until they were about to do evaluations. It would be
handy for them to actually use these while developing the program, and
while the ideal solution would be to have them read the whole
assignment, moving the test cases closer to the description of the
assignment rather than having it below the evaluation rubric might help.
A little. One hopes.