renamed created_at in decorator to created_at_representation and used it in show!
Overriding the created_at, created some issues with "distance_of_time_in_words(@article.created_at, comment.created_at)" part of the show view! 
Since we are changing the created_at definition when we're using it in the calculation of time distance with comment, the return value of the new created_at method is being used which is not a proper UTC time for that method and that's why the result of time distance of comment.created_at and @article.created_at which was 1 day for the first comment before this change now is becoming 8 months! (The same for the rest of the comments)

Thanks,