Add `reload!` to irb after adding `tag_list` method.
We were already working in the rails console a few lines up and then the tutorial has us change our model and then we jump back into the rails console. Without the `reload!` you will get a "no method" error which caused me to think I had a typo somewhere or I put my method in the wrong model, but all you have to do is reload the rails console.