Update typos in lines 1349-1353
Was, "Afterwards we actually send all the entire form letter content to the file object. The file object responds to the message puts. The file#puts is similar the Kernel#puts that we have been using up to this point."

is now, "Afterwards we actually send the entire form letter content to the file
object. The `file` object responds to the message `puts`. The
[file#puts](http://rubydoc.info/stdlib/core/IO#puts-instance_method) is similar to 
the [Kernel#puts](http://rubydoc.info/stdlib/core/Kernel#puts-instance_method)
that we have been using up to this point."