Added sentence for clarity.
I was confused that the Article.find(1) did not return the article's comments, but rather only the single article object. If you want to return the comments, you can do a = Article.find(1) which runs the database queries (to include the comments) followed by a.comments which returns all of that article's comments, but does not need to run another SQL query to do it.