Update eventmanager.markdown
Main purpose of this proposed change is to change all file paths from "../filename" to "filename" so that things actually run when the the file is run from the project root directory as demonstrated in the tutorial. This was mentioned in issue #936. Fixed a couple of typos I found along the way.

Also of note: on line 1046 and 1047, it says we will later move the form_letter.html (which later becomes form_letter.erb) to a new location and will be more explicit on defining the location of the template, but unless I'm missing it, this doesn't happen. I haven't changed anything there because I don't know whether it's preferable to leave it there and go ahead and change the location later (adding the relevant instructions to the tutorial) or to remove that part and let it be.