by_month
Sorry if this is a duplicate pull request. 

The by_month method is difficult to understand. Is this looking for merchants who have any sales in that month? Or is it looking for a ranking of sales by that month? Would really be helpful to understand how the output is generated

Thanks again - Allan