Make it clear what a migration is.
The paragraph following the header "What is a migration" does not explain what it is. Which seems rather odd considering the title of the paragraph. Personally I had to use third party resources to figure out what it even was. Therefore I think a short sentence explaining what it is is necesarry.

Olav